updated things

+++ wafbot/gamesnstuff.py
@@ -135,10 +135,12 @@ async def dm(msg):
 
 @client.event
 async def say(msg):
-    words = msg.content
-    words = words.replace("=say ", "")
-    await msg.delete()
-    await msg.channel.send(str(words))
+    if msg.author.bot is False:
+
+        words = msg.content
+        words = words.replace("=say ", "")
+        await msg.delete()
+        await msg.channel.send(str(words))
 
 @client.event
 async def afk(msg):
@@ -292,7 +294,7 @@ async def announce(msg):
 
 async def bigletters(msg):
     if msg.author.id == 698707989531459595:
-        letter = msg.content
+        letter = msg.content.lower()
         letters = letter.replace("=bigletters ", "")
         dictt = {'a':':regional_indicator_a:','b':':regional_indicator_b:','c':':regional_indicator_c:','d':':regional_indicator_d:','e':':regional_indicator_e:','f':':regional_indicator_f:','g':':regional_indicator_g:','h':':regional_indicator_h:',
         'i':':regional_indicator_i:','j':':regional_indicator_j:','k':':regional_indicator_k:','l':':regional_indicator_l:','m':':regional_indicator_m:','n':':regional_indicator_n:','o':':regional_indicator_o:','p':':regional_indicator_p:','q':':regional_indicator_q:',
